Overlapping code Changes : Files you need:
1. GenPermanence.cpp
2. GenPermanenceCalculation.h
3. OverlappingSeeding.h
4. Define Structures
   basic_comm.h
   defs.h
5. driverForGraphOverlappingClustering.cpp
6. Makefile
7. testgraph2.txt

+++ DefineStructure/defs.h
@@ -104,9 +104,9 @@ struct Perm_Info
 
 struct int2_dbl
 {
-    int first;
-    double second;
-    int third;
+    long first;
+    long second;
+    long third;
 
     //Constructor
     int2_dbl( )
